feat: 피드백 목록 조회 시 이모지 조회 로직 수정(#113)
