+++ src/main/java/reviewme/be/feedback/dto/response/FeedbackCommentResponse.java
@@ -1,17 +1,17 @@
 package reviewme.be.feedback.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.querydsl.core.annotations.QueryProjection;
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
 import lombok.Getter;
-import reviewme.be.feedback.dto.FeedbackCommentInfo;
+import lombok.Setter;
 import reviewme.be.util.dto.EmojiCount;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
-@Builder
+@Setter
 @Schema(description = "피드백 대댓글 목록 응답")
 public class FeedbackCommentResponse {
 
@@ -43,19 +43,17 @@ public class FeedbackCommentResponse {
     @Schema(description = "내가 선택한 이모지", example = "1")
     private Integer myEmojiId;
 
-    public static FeedbackCommentResponse fromFeedbackComment(FeedbackCommentInfo feedbackComment,
-        List<EmojiCount> emojis) {
+    @QueryProjection
+    public FeedbackCommentResponse(long id, long parentFeedbackId, String content, long commenterId, String commenterName,
+        String commenterProfileUrl, LocalDateTime createdAt, Integer myEmojiId) {
 
-        return FeedbackCommentResponse.builder()
-            .id(feedbackComment.getId())
-            .parentFeedbackId(feedbackComment.getParentId())
-            .content(feedbackComment.getContent())
-            .commenterId(feedbackComment.getCommenterId())
-            .commenterName(feedbackComment.getCommenterName())
-            .commenterProfileUrl(feedbackComment.getCommenterProfileUrl())
-            .createdAt(feedbackComment.getCreatedAt())
-            .emojis(emojis)
-            .myEmojiId(feedbackComment.getMyEmojiId())
-            .build();
+        this.id = id;
+        this.parentFeedbackId = parentFeedbackId;
+        this.content = content;
+        this.commenterId = commenterId;
+        this.commenterName = commenterName;
+        this.commenterProfileUrl = commenterProfileUrl;
+        this.createdAt = createdAt;
+        this.myEmojiId = myEmojiId;
     }
 }
