+++ src/main/java/reviewme/be/feedback/repository/FeedbackEmojiRepositoryCustom.java
@@ -5,5 +5,5 @@ import reviewme.be.util.dto.EmojiCount;
 
 public interface FeedbackEmojiRepositoryCustom {
 
-    List<EmojiCount> findEmojiCountByFeedbackIds(List<Long> feedbackIds);
+    List<EmojiCount> findFeedbackEmojiCountByFeedbackId(long feedbackId);
 }
